マテリアルのコマンドを整理

+++ Dev/ace_cpp/core/Graphics/Platform/DX11/Resource/ace.NativeShader_Imp_DX11.cpp
@@ -469,7 +469,7 @@ void NativeShader_Imp_DX11::SetConstantBuffer(const char* name, const void* data
 
 void NativeShader_Imp_DX11::SetConstantBuffer(int32_t id, const void* data, int32_t size)
 {
-	assert(id < constantLayoutsArray.size());
+	assert(id < (int32_t) constantLayoutsArray.size());
 	if (id < 0) return;
 
 	auto& layout = constantLayoutsArray[id];
@@ -514,7 +514,7 @@ void NativeShader_Imp_DX11::SetTexture(const char* name, Texture* texture, Textu
 
 void NativeShader_Imp_DX11::SetTexture(int32_t id, Texture* texture, TextureFilterType filterType, TextureWrapType wrapType)
 {
-	assert(id < textureLayoutsArray.size());
+	assert(id < (int32_t) textureLayoutsArray.size());
 	if (id < 0) return;
 
 	auto& layout = textureLayoutsArray[id];
